Move `qual_inspect` to a reusable location.
