+++ pg_search/src/postgres/customscan/mod.rs
@@ -29,12 +29,14 @@ use std::ptr::NonNull;
 mod builders;
 mod dsm;
 mod exec;
-mod hook;
-mod path;
-mod scan;
-
 mod explainer;
+mod hook;
+mod opexpr;
+mod path;
 pub mod pdbscan;
+mod pushdown;
+mod qual_inspect;
+mod scan;
 
 use crate::api::HashMap;
 use crate::postgres::customscan::exec::{
@@ -262,4 +264,14 @@ pub unsafe fn operator_oid(signature: &str) -> pg_sys::Oid {
         &[CString::new(signature).into_datum()],
     )
     .expect("should be able to lookup operator signature")
+}
+
+pub fn score_funcoid() -> pg_sys::Oid {
+    unsafe {
+        direct_function_call::<pg_sys::Oid>(
+            pg_sys::regprocedurein,
+            &[c"paradedb.score(anyelement)".into_datum()],
+        )
+        .expect("the `paradedb.score(anyelement)` function should exist")
+    }
 }
