+++ pg_search/src/postgres/customscan/pdbscan/exec_methods/fast_fields/mod.rs
@@ -28,8 +28,9 @@ use crate::nodecast;
 use crate::postgres::customscan::builders::custom_state::CustomScanStateWrapper;
 use crate::postgres::customscan::explainer::Explainer;
 use crate::postgres::customscan::pdbscan::privdat::PrivateData;
-use crate::postgres::customscan::pdbscan::projections::score::{score_funcoid, uses_scores};
+use crate::postgres::customscan::pdbscan::projections::score::uses_scores;
 use crate::postgres::customscan::pdbscan::{scan_state::PdbScanState, PdbScan};
+use crate::postgres::customscan::score_funcoid;
 use crate::postgres::rel::PgSearchRelation;
 use crate::schema::SearchIndexSchema;
 use itertools::Itertools;
